+++ controller/adminsController.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+router.post('/admins', (req, res) => {
+    res.render('admins');
+});
+
+
+module.exports = router;
