+++ controller/adminsWaitRoomController.js
@@ -1,0 +1,13 @@
+const express = require('express');
+const router = express.Router();
+
+router.post('/admins/wait', (req, res) => {
+    res.render('adminsWaitRoom', {
+        roomId: require('../src/idGenerator')(),
+        peopleNum: req.body.peopleNum
+    });
+
+});
+
+
+module.exports = router;
