+++ controller/participantsController.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+router.post('/participants', (req, res) => {
+    res.render('participants');
+});
+
+
+module.exports = router;
